deleted wtforms, added multiple html files, rerouted the document, added frontend folder, cleaned up application.py and added song input
