+++ frontend/templates/added_song.html
@@ -1,0 +1,6 @@
+{% extends "base2.html" %} {% block title %}Dodata pesma{% endblock %} {% block content
+    %}
+    <h1 align="center">Congratulations, song [kako dodati ime pesme?] is successfully added</h1>
+    <p>proceed to the Charts page</p>
+    
+{% endblock %}
