+++ frontend/templates/base.html
@@ -28,7 +28,7 @@
                     <a class="nav-item nav-link" id="login" href="login">Login</a>
                     <a class="nav-item nav-link" id="sign-up" href="sign-up">Sign Up</a>
                     <a class="nav-item nav-link" id="logout" href="logout">Logout</a>
-                    <a class="nav-item nav-link" id="Home" href="">Home</a>
+                    <a class="nav-item nav-link" id="Home" href="/">Home</a>
                 </div>
             </div>
         </nav>
