+++ frontend/templates/charts.html
@@ -1,0 +1,8 @@
+{% extends "base2.html" %} {% block title %}Charts{% endblock %} {% block content
+    %}
+    <h1 align="center">Charts</h1>
+    <p>kako ovo dizajnirati? i kakva je funkcionalnost?
+        Moja ideja je da se selektira Zanr i da se u zavisnosti od zanra prikaze top 10
+    </p>
+
+{% endblock %}
