+++ frontend/templates/sign_up.html
@@ -19,5 +19,8 @@
             <br/>
             <button type="submit" class= "btn btn-primary">Submit</button>
         </div>
+        <div>
+            <label id="label"> kako ovde staiviti token? </label>
+        </div>
     </form>
     {% endblock %}
