+++ frontend/templates/songs.html
@@ -1,0 +1,40 @@
+{% extends "base2.html" %} {% block title %}Songs{% endblock %} {% block content
+    %}
+    <h1 align="center">Songs</h1>
+    
+    <form method="POST">
+        <div class="form-group">
+            <label for="song">Song</label>
+            <input type="song" class=form-control id="somg" name="song" placeholder="Enter song"/>
+
+        </div>
+        
+        <div class="form-group">
+            <label for="artist">Artist</label>
+            <input type="artist" class=form-control id="artist" name="artist" placeholder="Enter Artist"/>
+
+        </div>
+        <div class="form-group">
+            <label for="genre">Genre</label>
+            <input type="genre" class=form-control id="genre" name="genre" placeholder="Enter Genre"/>
+
+        </div>
+        <div class="form-group">
+            <label for="yt_link">yt_link</label>
+            <input type="yt_link" class=form-control id="yt_link" name="yt_link" placeholder="Enter Youtube link"/>
+
+        </div>
+        <div class="form-group">
+            <label for="rating">Da li ovde treba rating?</label>
+            <input type="rating" class=form-control id="rating" name="rating" placeholder="Da li ovde treba rating?"/>
+
+        </div>
+
+        <div class="form-group">
+            <label for="token">token</label>
+            <input type="token" class=form-control id="token" name="token" placeholder="Enter Your token"/>
+
+        </div>
+        <button type="submit" class="btn btn-primary">Add Song</button>
+    </form>
+    {% endblock %}
