+++ frontend/templates/user.html
@@ -1,0 +1,10 @@
+{% extends "base2.html" %} {% block title %}User{% endblock %} {% block content
+    %}
+    %{with user=get_username}
+    <h1 align="center">Hello {{user}} [Kako ubaciti usera?] </h1>
+    <p>Your token for entering songs is[Kako ubaciti token?] {{token}} </p>
+    <p>Proceed to Songs page</p>
+
+
+
+{% endblock %}
