--- wtform/contact.html
@@ -1,52 +1,0 @@
-{% extends 'layout.html' %}
-
-
-{% block content %}
-
-  <div class="form-wrapper">
-    <h2 class="title">Contact</h2>
-
-    <form method="POST" action="{{ url_for('signup') }}">
-
-      <fieldset class="form-field">
-      {{ form.name.label }}
-      {{ form.name }}
-        {% if form.name.errors %}
-        <ul class="errors">
-          {% for error in form.name.errors %}
-            <li>{{ error }}</li>
-          {% endfor %}
-        </ul>
-        {% endif %}
-      </fieldset>
-
-      <fieldset class="form-field">
-        {{ form.password.label }}
-        {{ form.password }}
-        {% if form.password.errors %}
-          <ul class="errors">
-            {% for error in form.password.errors %}
-              <li>{{ error }}</li>
-            {% endfor %}
-          </ul>
-        {% endif %}
-      </fieldset>
-
-      <fieldset class="form-field">
-        {{ form.song.label }}
-        {{ form.song }}
-        {% if form.song.errors %}
-          <ul class="errors">
-            {% for error in form.song.errors %}
-              <li>{{ error }}</li>
-            {% endfor %}
-          </ul>
-        {% endif %}
-      </fieldset>
-
-      {{ form.submit }}
-
-      </form>
-  </div>
-
-{% endblock %}
