--- wtform/contact.jinja2
@@ -1,31 +1,0 @@
-{% extends 'layout.html' %}
-
-
-{% block content %}
-
-    <div class="form-wrapper">
-      <h2 class="title">Contact</h2>
-
-      <form method="POST" action="{{ url_for('contact') }}">
-
-          <fieldset class="form-field">
-              {{ form.name.label }}
-              {{ form.name(size=20) }}
-          </fieldset>
-
-          <fieldset class="form-field">
-              {{ form.password.label }}
-              {{ form.password }}
-          </fieldset>
-
-          <fieldset class="form-field">
-              {{ form.song.label }}
-              {{ form.song }}
-          </fieldset>
-
-          {{ form.submit }}
-
-      </form>
-    </div>
-
-{% endblock %}
